// packages/cdk/src/network/virtual-network-types.d.ts
export interface AddressSpace {
    /**
     * Array of IP address ranges in CIDR notation.
     *
     * @example ['10.0.0.0/16', '10.1.0.0/16']
     */
    readonly addressPrefixes: string[];
}
/**
 * DHCP options for virtual network.
 */
export interface DhcpOptions {
    /**
     * Array of DNS server IP addresses.
     *
     * @example ['10.0.0.4', '10.0.0.5']
     */
    readonly dnsServers: string[];
}
// ...
export interface InlineSubnetProps {
    /**
     * Name of the subnet.
     */
    readonly name: string;
    /**
     * Address prefix for the subnet in CIDR notation.
     *
     * @remarks
     * Must be within the VNet's address space.
     *
     * @example '10.0.1.0/24'
     */
    readonly addressPrefix: string;
    /**
     * Network security group ID to associate with the subnet.
     *
     * @example '/subscriptions/{sub}/resourceGroups/{rg}/providers/Microsoft.Network/networkSecurityGroups/{nsg}'
     */
    readonly networkSecurityGroup?: {
        id: string;
    };
    /**
     * Service endpoints for the subnet.
     *
     * @example [{ service: 'Microsoft.Storage' }, { service: 'Microsoft.Sql' }]
     */
    readonly serviceEndpoints?: Array<{
        service: string;
    }>;
    /**
     * Delegations for the subnet.
     *
     * @example [{ name: 'delegation', serviceName: 'Microsoft.Web/serverFarms' }]
     */
    readonly delegations?: Array<{
        name: string;
        serviceName: string;
    }>;
}
// ...
export interface ArmVirtualNetworkProps {
    /**
     * Name of the virtual network.
     */
    readonly virtualNetworkName: string;
    /**
     * Azure region where the virtual network will be created.
     */
    readonly location: string;
    /**
     * Resource group name where the VNet will be deployed.
     *
     * @remarks
     * Required for generating the full resource ID.
     */
    readonly resourceGroupName: string;
    /**
     * Address space for the virtual network.
     *
     * @remarks
     * Must contain at least one address prefix in CIDR notation.
     */
    readonly addressSpace: AddressSpace;
    /**
     * Subnets to create within the virtual network.
     *
     * @remarks
     * Can be defined inline or created separately after VNet deployment.
     */
    readonly subnets?: InlineSubnetProps[];
    /**
     * DHCP options (DNS servers).
     */
    readonly dhcpOptions?: DhcpOptions;
    /**
     * Enable DDoS protection.
     *
     * @default false
     */
    readonly enableDdosProtection?: boolean;
    /**
     * Enable VM protection.
     *
     * @default false
     */
    readonly enableVmProtection?: boolean;
    /**
     * Tags to apply to the virtual network.
     */
    readonly tags?: Record<string, string>;
}